[GEOS-10735] Obfuscate secret key in S3 Blob Store, avoiding requiring reentry when editing and HTML source visibility
cleanup

test cleanup
